Add stateful dirty set to keep track across runs
shakeRunDatabaseForKeys is an unsound approximation since it bypasses the
version number checks. It is possible that a key K is not up to date and yet,
because it doesn't belong to the dirty set, gets reused without updating.

In what cases can this happen? The most common one is when a previous call to
shakeRunDatabaseForKeys marked K dirty, but the evaluation didn't touch K so it
didn't get recomputed. The current call to shakeRunDatabaseForKeys doesn't know
this, and if the new dirty set doesn't include K but the evaluation uses it, the
stale value will be reused.

To avoid this, we persist (in memory only) the dirty set and remember it across
runs. A key is removed from the persisted dirty set as it is refreshed.